requirements.txt
moviepy
SpeechRecognition
openai-whisper